Renamed CommsCell and OutCurr_Cell to make the classes less ambiguous
